Move sanatize.py to utils module
Update typing and docstrings
Correct typo sanatize -> sanitize (incl rename to sanitize.py)
